+++ sphinx/source/changelog.rst
@@ -5,6 +5,8 @@ Changelog
 -----
 * Added the ``--call-graph`` option to :ref:`snapinfo.py` (for generating a
   call graph in DOT format)
+* Added the ``--ctl`` option to :ref:`snapinfo.py` (for specifying a control
+  file to use when generating a call graph)
 * Added support to :ref:`snapinfo.py <snapinfo-conf>` for reading configuration
   from `skoolkit.ini`
 * Added the ``--ini`` and ``--show-config`` options to :ref:`snapinfo.py` (for